requirements.txt
# Core Django
Django==5.2.0
psycopg2-binary==2.9.11
django-environ==0.11.2

# Django Rest Framework
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1

# Telegram bot
python-telegram-bot==20.7

# Celery & Redis
celery==5.3.4
redis==5.0.1

# Testing
pytest==7.4.3
pytest-asyncio==1.3.0
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
Faker==22.0.0

# Code Quality
black==23.12.1
isort==5.13.2
flake8==7.0.0
pre-commit==3.6.0

# Utilities
python-decouple==3.8
django-extensions==3.2.3
ipython==8.19.0

# API Documentation
drf-spectacular==0.27.0
